requirements.txt
psycopg2-binary>=2.9.3
akshare>=1.3.0
pandas>=1.3.0
pandas-market-calendars>=3.3
python-dotenv>=0.19.0
pytest>=7.0
pytest-cov>=3.0
pytest-mock>=3.0
freezegun>=1.0  # 用于时间相关测试